requirements.txt
celery==5.1.2
redis==3.5.3
django-celery-results==2.2.0
Django>=3.0,<4.0
psycopg2-binary>=2.8
celery-progress==0.1.1
django-crispy-forms==1.12.0